backend/requirements_simple.txt
# Simplified Backend Requirements
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.5.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.1
openai>=1.3.0
requests>=2.31.0
supabase==2.17.0
